% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_springer_book.R
\name{download_springer_book}
\alias{download_springer_book}
\title{Function that fetchs the pdf file and saves it in the current directory}
\usage{
download_springer_book(book_spec_title, springer_table)
}
\description{
Function that fetchs the pdf file and saves it in the current directory
}
