% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_springer_book_files.R
\name{download_springer_book_files}
\alias{download_springer_book_files}
\title{Generate all pdf's in the directories organized by book type}
\usage{
download_springer_book_files(
  springer_books_titles = NA,
  springer_table = NA,
  destination_folder = "springer_quarantine_books"
)
}
\description{
Generate all pdf's in the directories organized by book type
}
