% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/springer_book_categories.R
\name{springer_book_categories}
\alias{springer_book_categories}
\title{List of springer book categories}
\usage{
springer_book_categories()
}
\value{
The return of this function is a tibble containing Springer categories
}
\description{
\code{springer_book_categories} This function returns a tibble that contains the categories used by springer to classify its books.
It can be used to filter book categories to download in `download_springer_book_files()` function.
}
\details{
This is a void function, no arguments are passed to it
}
\examples{

\dontrun{
springer_book_categories()
}

}
