% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_springer_book_files.R
\name{download_springer_book_files}
\alias{download_springer_book_files}
\title{Downloads multiple springer books, allowing filters of book groups}
\usage{
download_springer_book_files(
  springer_books_titles = NULL,
  springer_book_group = NULL,
  springer_table = NULL,
  destination_folder = "books",
  parallel = FALSE
)
}
\arguments{
\item{springer_books_titles}{A list of books to be downloaded, if left empty, it'll download every book.}

\item{springer_book_group}{A list of groups which to download books from (the default list can be found at `springer_book_categories()`),
if left empty, it'll download books from every category}

\item{springer_table}{The default table exported from springer website, if left empty, will source the provided table in web.}

\item{destination_folder}{A folder/path that will be used to save the files.}

\item{parallel}{A boolean which the user can choose to run the supplied function in parallel or not.}
}
\value{
The function will download the pdf of the books listed/provided, export it at the provided destination folder, and there are no
returns for this function
}
\description{
\code{download_springer_book_files} This function can receive many arguments that will be used to download books from
the Springer open repository, it can receive a list of the names of the books that the user want to download,
it can receive a specific group of books to be downloaded, such as engineering, psycology, etc. (the list of available categories
can be found by running the `springer_book_categories()` function), if provided, it'll use a specific springer table to source
the url and book specifications, it also can receive a destination folder, and an argument that will allow the user to
run this in parallel through future.
}
\examples{

\dontrun{
download_springer_book_files()
}

}
