% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_springer_book_files.R
\name{download_springer_book_files}
\alias{download_springer_book_files}
\title{Download all pdf's/epub's in the directories organized by book type}
\usage{
download_springer_book_files(
  springer_books_titles = NA,
  destination_folder = "springer_quarantine_books",
  lan = "eng",
  filetype = "pdf"
)
}
\arguments{
\item{springer_books_titles}{A list of books to be downloaded, if left empty, it'll download every book.}

\item{destination_folder}{A folder/path that will be used to save the files.}

\item{lan}{The language of the downloaded books. Can be either set to 'eng' (English) or 'ger' (German). Default is 'eng'.}

\item{filetype}{The file type extension of the books downloaded. Can be either set to 'pdf', 'epub' or 'both'. Default is 'pdf'.}
}
\description{
\code{download_springer_book_files} This function can receive many arguments that will be used to download books from
the Springer open repository.
}
