% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_springer_table.R
\name{download_springer_table}
\alias{download_springer_table}
\title{Download the springer table file}
\usage{
download_springer_table(lan = "eng")
}
\arguments{
\item{lan}{The language of the downloaded books. Can be either set to 'eng' (English) or 'ger' (German). Default is 'eng'.}
}
\description{
\code{download_springer_table} This function downloads the springer table of all books available.
}
